fix: consolidate session save into factory_editable_definition
Merge factory_session_save.go and session save tests into existing
pkg/service files to satisfy the 15-file package limit after rebase.
