+++ docs/internal/processes/factory-workstation-relevant-files.md
@@ -134,7 +134,7 @@ work inputs.
 - Keep editable runner dropdown options on the generated `RunnerID` set from `OPENAPI_RUNNER_IDS` in `ui/src/features/current-selection/workstation-selection/messages/runner-openapi-enums.ts`, mirror backend precedence through `resolveRunnerSelection(...)` in `ui/src/features/current-factory-definition/lib/runner-selection.ts`, and render summary/form source labels via `localizeRunnerSelectionSource(...)` instead of parallel summary-only inheritance strings.
 - Repeater-aware rejected-outcome copy in `ui/src/features/current-selection/base/messages/current-selection-operational-enums.ts` should treat canonical `REPEATER` and legacy lowercase `repeater` workstation kinds the same when labeling `REJECTED` provider-session outcomes as repeated work; provider-session and dispatch history surfaces pass through `selectedNode.workstation_kind` unchanged.
 - When the dashboard saves current factory edits, use the canonical `PUT /factory-sessions/{session_id}/factory` route with `SaveFactoryForSessionRequest` (`factory` required, optional `mode` defaulting to `REPLACE_CURRENT`) and keep the typed current-factory adapter local to `ui/src/api/current-factory-definition/api.ts`; `saveCurrentFactoryDocument` and `useSaveCurrentFactory` always send explicit `mode: REPLACE_CURRENT` via `CURRENT_FACTORY_EDITOR_SAVE_MODE` with incremented `factory.version` on the payload—editor save paths must not call `UPSERT_NAMED_AND_ACTIVATE`. Dashboard import confirm dialog offers replace-current vs create-new-named; create-new uses `activateImportedFactoryAsNewNamedForSession` with suffixed naming from `ui/src/api/named-factory/import-save-mode.ts`. Keep mutation orchestration in `ui/src/features/current-selection/` and update both seams on success: write the normalized saved `Factory` document into the `current-editable-factory-definition` React Query cache and mark the local workstation draft clean in the same success path so the header save action, form status, and saved values stay aligned without a second local shadow copy.
-- Backend factory submission is unified in `pkg/service/factory_session_save.go` `SaveFactoryForSession` (resolve session scope → validate → persist by mode → `replaceSessionRuntime` → readback). `REPLACE_CURRENT` uses `factorysessions.SessionFactoryRootDir` and always requires an advanced `factory.version` via `requireFreshEditableFactoryVersionAtRoot`; `UPSERT_NAMED_AND_ACTIVATE` persists under `sessionFactoryPersistRoot` (non-default sessions use `LiveSession.FolderPath`, default session uses the service root), omits version on create, and uses on-disk version for stale detection on replace. Handler `SaveCurrentFactoryBySessionId` in `pkg/api/handlers.go` calls only `SaveFactoryForSession`; legacy `CreateNamedFactory`, unscoped `SaveCurrentFactory`, and save-path `activateReplacementRuntime` are removed. `GetCurrentFactoryForSession` prefers `current-factory` pointer names from `sessionFactoryPersistRoot` before `factorysessions.FactoryName` derivation. Prove per-mode version rules in `pkg/service/factory_session_save_version_test.go` and `tests/functional/runtime_api/factory_transformation/api_session_factory_save_version_test.go`; prove non-default session isolation with `TestFactoryService_SaveFactoryForSession_UpsertOnNonDefaultSessionDoesNotMutateDefault` and `TestFactoryService_SaveCurrentFactoryForSession_ReplacesOnlyTargetedSession`.
+- Backend factory submission is unified in `pkg/service/factory_editable_definition.go` (`SaveFactoryForSession` and mode handlers) `SaveFactoryForSession` (resolve session scope → validate → persist by mode → `replaceSessionRuntime` → readback). `REPLACE_CURRENT` uses `factorysessions.SessionFactoryRootDir` and always requires an advanced `factory.version` via `requireFreshEditableFactoryVersionAtRoot`; `UPSERT_NAMED_AND_ACTIVATE` persists under `sessionFactoryPersistRoot` (non-default sessions use `LiveSession.FolderPath`, default session uses the service root), omits version on create, and uses on-disk version for stale detection on replace. Handler `SaveCurrentFactoryBySessionId` in `pkg/api/handlers.go` calls only `SaveFactoryForSession`; legacy `CreateNamedFactory`, unscoped `SaveCurrentFactory`, and save-path `activateReplacementRuntime` are removed. `GetCurrentFactoryForSession` prefers `current-factory` pointer names from `sessionFactoryPersistRoot` before `factorysessions.FactoryName` derivation. Prove per-mode version rules in `pkg/service/factory_session_save_version_test.go` and `tests/functional/runtime_api/factory_transformation/api_session_factory_save_version_test.go`; prove non-default session isolation with `TestFactoryService_SaveFactoryForSession_UpsertOnNonDefaultSessionDoesNotMutateDefault` and `TestFactoryService_SaveCurrentFactoryForSession_ReplacesOnlyTargetedSession`.
 - When editable-definition saves reject topology before persistence, return `apisurface.TopologyValidationError` targets that map graph issues to `node`, `edge`, `field`, or `save` locations; the API handler should preserve those targets in the `INVALID_FACTORY` response instead of collapsing every validation failure to the form root.
 - When the current-selection workstation editor narrows to workstation-scoped fields such as `worker` and `prompt`, keep the mutation path anchored on the canonical current-factory definition but only apply the fields that truly belong to the selected workstation; preserve unrelated worker-model and template data from the latest fetched definition instead of carrying stale hidden values forward into the editable-definition save payload.
 - When a workstation edit surface exposes worker-backed fields such as `model`, verify whether the referenced worker is shared by sibling workstations before treating the field as workstation-scoped; if workers can be shared, disable or clearly rescope that edit so one workstation save cannot silently rewrite others.
